requirements.txt
absl-py==1.2.0
astunparse==1.6.3
autopep8==1.6.0
# box2d-py==2.3.5
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
cycler==0.11.0
flatbuffers==1.12
fonttools==4.34.4
gast==0.4.0
google-auth==2.10.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
gym==0.25.1
gym-notices==0.0.7
h5py==3.7.0
idna==3.3
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
MarkupSafe==2.1.1
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
Pillow==9.2.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pygame==2.1.0
pyglet==1.5.26
pyparsing==3.0.9
python-dateutil==2.8.2
requests-oauthlib==1.3.1
rsa==4.9
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
toml==0.10.2
typing_extensions==4.3.0
urllib3==1.26.11
Werkzeug==2.2.1
wrapt==1.14.1
black~=22.3.0
pygments>=2.7.4
markdown>=3.3
matplotlib
jinja2
requests
mdx_truly_sane_lists
PyYAML
pandas
jupyter
tensorflow~=2.11.0
tensorflow_datasets
keras-tuner
keras-cv
keras-nlp
pycocotools